Removed old example datasets, renamed some files
