--- DashAI/back/example_dataset/dummy_numeric.json
@@ -1,113 +1,0 @@
-{
-    "task_info": {
-        "task_type": "TabularClassificationTask"
-    },
-    "train": {
-        "x": [
-            [
-                1,
-                2,
-                3
-            ],
-            [
-                2,
-                3,
-                4
-            ],
-            [
-                3,
-                2,
-                1
-            ],
-            [
-                5,
-                2,
-                0
-            ],
-            [
-                1,
-                2,
-                3
-            ],
-            [
-                2,
-                3,
-                4
-            ],
-            [
-                3,
-                2,
-                1
-            ],
-            [
-                5,
-                2,
-                0
-            ]
-        ],
-        "y": [
-            "1",
-            "0",
-            "1",
-            "1",
-            "1",
-            "0",
-            "1",
-            "1"
-        ]
-    },
-    "test": {
-        "x": [
-            [
-                1,
-                2,
-                3
-            ],
-            [
-                2,
-                3,
-                4
-            ],
-            [
-                3,
-                2,
-                1
-            ],
-            [
-                5,
-                2,
-                0
-            ],
-            [
-                1,
-                2,
-                3
-            ],
-            [
-                2,
-                3,
-                4
-            ],
-            [
-                3,
-                2,
-                1
-            ],
-            [
-                5,
-                2,
-                0
-            ]
-        ],
-        "y": [
-            "1",
-            "0",
-            "1",
-            "1",
-            "1",
-            "0",
-            "1",
-            "1"
-        ]
-    }
-}
