--- DashAI/back/example_dataset/dummy_text.json
@@ -1,34 +1,0 @@
-{
-    "name": "Dummy Text Dataset",
-    "task_info":{
-        "task_type":"TextClassificationTask"
-    },
-    "train":{
-        "x":[
-            "hola",
-            "como",
-            "estas",
-            "tu"
-        ],
-        "y":[
-            "1",
-            "0",
-            "1",
-            "1"
-        ]
-    },
-    "test":{
-        "x":[
-            "hola",
-            "como",
-            "estas",
-            "tu"
-        ],
-        "y":[
-            "1",
-            "0",
-            "1",
-            "1"
-        ]
-    }
-}
