+++ DashAI/back/routers/session_class.py
@@ -5,5 +5,8 @@ class Session(object):
     task_name = ""
     task = None
 
+    def state(self):
+        return {"dataset": self.dataset, "task_name": self.task_name, "task": self.task}
+
 
 session_info = Session()  # TODO Remove this
